refactor: move all protocols imports to a single file
