+++ src/presentation/controllers/signup.spec.ts
@@ -1,6 +1,5 @@
 import { MissingParamError, InvalidParamError, ServerError } from '../errors'
-import { EmailValidator } from '../protocols/email-validator'
-import { HttpRequest, HttpResponse } from '../protocols/http'
+import { HttpRequest, HttpResponse, EmailValidator } from '../protocols'
 import { SignUpController } from './signup'
 
 interface SutTypes {
